Modularize knowledge workflow
